+++ crates/jsonschema/src/lib.rs
@@ -516,7 +516,7 @@ pub use error::{ErrorIterator, MaskedValidationError, ValidationError};
 pub use keywords::custom::Keyword;
 pub use options::ValidationOptions;
 pub use output::BasicOutput;
-pub use referencing::{Draft, Error as ReferencingError, Resource, Retrieve, Uri};
+pub use referencing::{Draft, Error as ReferencingError, Registry, Resource, Retrieve, Uri};
 pub use validator::Validator;
 
 use serde_json::Value;